refactor: moved ApiService.kt to di.services.network
